Update LICENSE to BSD-3-clause
GNU was used due to oversight and miscommunication. It was never the intent to restrict the use of this package, and the funding mechanism supporting this development mandates distribution under a non-restrictive open source license. See https://github.com/ImagingDataCommons/IDC-ProjectManagement/issues/1600. BSD-3-clause was chosen due to the precedent of using that license for pyradiomics.